move api config to the api folder
