+++ internal/api/impl/v1/dashboard/schemas/hotreload.go
@@ -20,10 +20,9 @@ import (
 	"strings"
 
 	"github.com/perses/common/async"
-	"github.com/perses/perses/internal/config"
+	"github.com/perses/perses/internal/api/config"
 	"github.com/sirupsen/logrus"
 	log "github.com/sirupsen/logrus"
-	"gopkg.in/fsnotify.v1"
 )
 
 type watcher struct {
