+++ internal/api/impl/v1/dashboard/service.go
@@ -17,11 +17,11 @@ import (
 	"fmt"
 
 	"github.com/perses/common/etcd"
+	"github.com/perses/perses/internal/api/config"
 	"github.com/perses/perses/internal/api/impl/v1/dashboard/schemas"
 	"github.com/perses/perses/internal/api/impl/v1/dashboard/variable"
 	"github.com/perses/perses/internal/api/interface/v1/dashboard"
 	"github.com/perses/perses/internal/api/shared"
-	"github.com/perses/perses/internal/config"
 	"github.com/perses/perses/pkg/model/api"
 	v1 "github.com/perses/perses/pkg/model/api/v1"
 	"github.com/sirupsen/logrus"
