+++ internal/api/shared/database/dao.go
@@ -24,8 +24,7 @@ import (
 	"time"
 
 	"github.com/perses/common/etcd"
-	"github.com/perses/perses/internal/config"
-	"gopkg.in/yaml.v2"
+	"github.com/perses/perses/internal/api/config"
 )
 
 type DAO interface {
