+++ internal/api/shared/dependency/persistence_manager.go
@@ -14,6 +14,7 @@
 package dependency
 
 import (
+	"github.com/perses/perses/internal/api/config"
 	dashboardImpl "github.com/perses/perses/internal/api/impl/v1/dashboard"
 	datasourceImpl "github.com/perses/perses/internal/api/impl/v1/datasource"
 	folderImpl "github.com/perses/perses/internal/api/impl/v1/folder"
@@ -29,7 +30,6 @@ import (
 	"github.com/perses/perses/internal/api/interface/v1/project"
 	"github.com/perses/perses/internal/api/interface/v1/user"
 	"github.com/perses/perses/internal/api/shared/database"
-	"github.com/perses/perses/internal/config"
 )
 
 type PersistenceManager interface {
