+++ internal/api/shared/dependency/service_manager.go
@@ -16,6 +16,7 @@
 package dependency
 
 import (
+	"github.com/perses/perses/internal/api/config"
 	dashboardImpl "github.com/perses/perses/internal/api/impl/v1/dashboard"
 	datasourceImpl "github.com/perses/perses/internal/api/impl/v1/datasource"
 	folderImpl "github.com/perses/perses/internal/api/impl/v1/folder"
@@ -30,7 +31,6 @@ import (
 	"github.com/perses/perses/internal/api/interface/v1/health"
 	"github.com/perses/perses/internal/api/interface/v1/project"
 	"github.com/perses/perses/internal/api/interface/v1/user"
-	"github.com/perses/perses/internal/config"
 )
 
 type ServiceManager interface {
